% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elev.R
\name{elev_url}
\alias{elev_url}
\title{Construct Data URL}
\usage{
elev_url(coords)
}
\arguments{
\item{coords}{\verb{[numeric(2) or data.frame]} Numeric vector of length 2 giving
the coordinates of a single point that falls in the desired tile or a
two-column data.frame giving the longitude and latitude (in that order) of
multiple points to be considered.}
}
\description{
Constructs the URL for the elevation data requested
}
\examples{

# Example location (near Bear Lake, UT)
coords1 <- c(-111.419, 41.94)

url1 <- elev_url(coords1)

# Example data.frame
coord_df <- data.frame(x = c(-111.5, -112.5),
                       y = c(41.5, 40.5))

url2 <- elev_url(coord_df)

}
